Added Trait Test, nonce test, and code refactor
