+++ tests/Models/TokenSigningTests.php
@@ -8,22 +8,63 @@
 namespace VanillaTests\Models;
 
 use VanillaTests\SharedBootstrapTestCase;
-use AccessTokenModel;
 use VanillaTests\SiteTestTrait;
-use Vanilla\TokenSigningTrait;
-use VanillaTests\Fixtures\TokenModel
+use VanillaTests\Fixtures\TokenModel;
 
 /**
- * Test the {@link AccessTokenModel}.
+ * Test the {@link TokenModel}.
  */
-class AccessTokenModelTest extends SharedBootstrapTestCase {
+class TokenSigningTests extends SharedBootstrapTestCase {
     use SiteTestTrait;
 
-    public function testVerifyToken () {
+    public function testVerifyRandomTokenSignature() {
         $model = new TokenModel();
+        $model->tokenIdentifier ='nonce';
+        $token = $model->randomSignedToken();
 
+        $this->assertEquals(true, $model->verifyTokenSignature($token, $model->tokenIdentifier));
     }
 
-    public function __construct() {
+    /**
+     * An expired token shouldn't verify.
+     *
+     * @expectedException \Exception
+     * @expectedExceptionMessage Your nonce has expired.
+     */
+    public function testExpiryDate() {
+        $model = new TokenModel();
+        $model->tokenIdentifier ='nonce';
+        $token = $model->randomSignedToken('last month');
+
+        $model->verifyTokenSignature($token, $model->tokenIdentifier, true);
+    }
+
+    /**
+     * An altered token signature shouldn't verify.
+     *
+     * @expectedException \Exception
+     * $expectedExceptionMessage Invalid signature.
+     */
+    public function testBadSignature() {
+        $model = new TokenModel();
+        $model->tokenIdentifier ='nonce';
+
+        $token = $model->randomSignedToken().'!';
+        $model->verifyTokenSignature($token, $model->tokenIdentifier, true);
+    }
+
+    /**
+     * A nonsense token shouldn't verify.
+     *
+     * @expectedException \Exception
+     */
+    public function testBadToken() {
+        $model = new TokenModel();
+        $model->tokenIdentifier = 'nonce';
+
+        $token = 'a.b.c';
+        $model->verifyTokenSignature($token, $model->tokenIdentifier, true);
+    }
+
 
 }
